+++ tests/integration.rs
@@ -12,34 +12,39 @@ use qrlew::{
     io::{postgresql, Database},
     sql::parse,
     Relation, With,
+    expr
 };
+
+pub fn test_eq<D: Database>(database: &mut D, query1: &str, query2: &str) -> bool {
+    println!(
+        "{}\n{}",
+        format!("{query1}").red(),
+        database
+            .query(query1)
+            .unwrap()
+            .iter()
+            .map(ToString::to_string)
+            .join("\n")
+    );
+    println!(
+        "{}\n{}",
+        format!("{query2}").yellow(),
+        database
+            .query(query2)
+            .unwrap()
+            .iter()
+            .map(ToString::to_string)
+            .join("\n")
+    );
+    database.eq(query1, query2)
+}
 
 pub fn test_rewritten_eq<D: Database>(database: &mut D, query: &str) -> bool {
     let relations = database.relations();
     let relation = Relation::try_from(parse(query).unwrap().with(&relations)).unwrap();
     let rewriten_query: &str = &ast::Query::from(&relation).to_string();
     relation.display_dot().unwrap();
-    println!(
-        "{}\n{}",
-        format!("{query}").red(),
-        database
-            .query(query)
-            .unwrap()
-            .iter()
-            .map(ToString::to_string)
-            .join("\n")
-    );
-    println!("{}", format!("{rewriten_query}").yellow());
-    println!(
-        "{}",
-        database
-            .query(rewriten_query)
-            .unwrap()
-            .iter()
-            .map(ToString::to_string)
-            .join("\n")
-    );
-    database.eq(query, rewriten_query)
+    test_eq(database, query, rewriten_query)
 }
 
 const QUERIES: &[&str] = &[
@@ -121,3 +126,33 @@ fn test_on_postgresql() {
         assert!(test_rewritten_eq(&mut database, query));
     }
 }
+
+#[test]
+fn test_distinct_aggregates() {
+    let mut database = postgresql::test_database();
+    let table = database.relations().get(&["table_1".to_string()]).unwrap().as_ref().clone();
+
+    let true_query = "SELECT COUNT(DISTINCT d) AS count_d, SUM(DISTINCT d) AS sum_d FROM table_1";
+    let column = "d";
+    let group_by = vec![];
+    let aggregates = vec![
+            ("count_d", expr::aggregate::Aggregate::Count),
+            ("sum_d",  expr::aggregate::Aggregate::Sum),
+    ];
+    let distinct_rel = table.clone().distinct_aggregates(column, group_by, aggregates);
+    let rewriten_query:&str = &ast::Query::from(&distinct_rel).to_string();
+    assert!(test_eq(&mut database, true_query, rewriten_query));
+
+    let true_query = "SELECT c, COUNT(DISTINCT d) AS count_d, SUM(DISTINCT d) AS sum_d FROM table_1 GROUP BY c ORDER BY c";
+    let column = "d";
+    let group_by = vec!["c"];
+    let aggregates = vec![
+            ("count_d", expr::aggregate::Aggregate::Count),
+            ("sum_d",  expr::aggregate::Aggregate::Sum),
+        ];
+    let distinct_rel = table.distinct_aggregates(column, group_by, aggregates);
+    let rewriten_query:&str = &ast::Query::from(&distinct_rel).to_string();
+    assert!(test_eq(&mut database, true_query, rewriten_query));
+}
+
+